Updated to version 2.3.0
WARNING: Code is still stupid and bad. Just added a fields list to help out some folks
-----
Added a dynamic Field list that you can add stuff to.
Set defaults in the Start Node per conversation.
Defaults will be added as nodes are viewed and will be added for the entire conversation on save.
Defaults can be overwritten per node.